web_portal/requirements.txt
asgiref==3.9.0
attrs==25.3.0
beautifulsoup4==4.12.3
celery==5.4.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
django-google-maps==0.13.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
hdbcli==2.26.25
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==5.3.0
mammoth==1.8.0
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
reportlab==4.2.5
requests==2.32.4
rpds-py==0.26.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
twilio==9.3.7
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
